feat(*): Completes spike of HA stream architecture
This adds some code for handling the event streams and a basic setup of
the wadm binary. Much of the code in the binary will likely need to be
refactored into library code, but it is good enough to prove that things
work. I might do another follow on PR to this to maybe clean up a few more
things.

Closes #42

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>