ref(storage): Refactors store to better fit the data structure we need
This is part of my work for actually generating the lattice state, however
I realized that in the previous task we didn't really specify how we were
going to save the data. As I started to work on storing lattice state, it
became obvious what we had wasn't a good fit. So I refactored to better
fit what we needed and took a little bit of time to make the patterns match
what we had in other places (such as taking a `Store`, rather than
constructing the client and the bucket within the store)

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>