feat(api): Implements all model operations
This PR implements the wadm NATS api and backend storage for the manifests.
Anywhere I have deviated from the current wadm_0.3 api, I have tried to
document it with a note so we can update the docs later. In a follow up PR,
I will finish implementing the actual deployment operations of the API. For
now, I wanted to make sure this didn't get _too_ big to review

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>