feat(*): Automatically observe lattices
This commit makes it so that wadm automatically watches for new lattices
to appear and then manages them. As part of this, it made sense to actually
make sure this is mutlitenant ready as well. So this implements the
functionality to properly observe for multitenant (read: cross account importing)
architectures.

Not gonna lie that some of the stuff to start watching pre-observed lattices is
a bit messy, but I did test it out and it is real slick

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>