feat(*): Notifies on manifest deployment
Please note that this is a bit big because I had to have a stream that
combined internal wadm events with wasmbus ones. However, I did realize
that we only care about streams for their configuration and delivery
guarantees, and not for their storage capabilities. Because of this,
we don't actually have to have the users configure the streams beforehand.
So while this adds one more stream to manage combining the two topics
(until nats 2.10 comes out and we move to it), the user no longer
has to configure any of them.

Closes #49

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>