feat(*): Integrates in the scalers into the event workers
This is a giant PR because a lot of debugging happened. Most of this stuff
is Real Code we will keep around. However, this is one big hacky
thing to point out. In order to handle Actor Start/Stop events, we had
to introduce the idea of a backoff, similar to what is done in wadm 0.3.
This method definitely puts the "eventual" in eventual consistency,
particularly with high replica counts, but it does work. This backoff
is not needed for linkdef or provider scalers. The reason why this was
needed was due to there being no way for identifying "groups" (for lack
of a better term) of started and stopped events. We ended up with a
stream of events with a non-deterministic end. The real solution here
would be to update our actor started/stopped events to just emit a
single event when a requested scale is done. So if a command comes in
to start 10 actors, then we get all the same stuff we current do with
a new `count` field and sans the instance-id field, which I don't think
we use anywhere except the host and lattice observer (and I think the
observer can do without it). This is also better for state storage as
we don't have to read and write for _every_ single actor that starts
or stops

Closes #49

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>