feat(*): Finishes up work for distributed backoff
This fixes several minor issues and streamlines how backoffs are handled.

First, to get around issues with event serialization when resending the
event as part of a notification, we now reencode it as a cloudevent. There
was a lot of work to be done to manage our own serialization with it.

Second, this moves all backoff code to be internal to the backoff scaler
wrapper. This helps avoid leaky abstractions in the future and made the
code easier to debug

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>