fix(*): Link caching and other goodies
This PR got a little bit out of hand. We had a bunch of bugs that may or
may not have been interrelated so we fixed most of them. The biggest thing
is that we switching to caching link definitions for scalers so that
the number of consumers created was much lower. Additionally, we fixed when
status notifications are published so that we don't get reconciling status
when other scalers are being removed

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>