fix: increase length to 512 for image references
Sometimes, sufficiently nested folders will require a length longer
than 128 characters for the path to a Wasm executable, when using the
`file://` specifier.

While it's questionable whether the image should be limited at all, it
looks like 128 is at the very least too low.

This commit increases the string allowed for `image` specification to
a higher multiple of 128.

Signed-off-by: Victor Adossi <vadossi@cosmonic.com>