feat: support TLS CA certificates when building a NATS client
NATS servers support TLS, which means sometimes you need to be able to
provide certificates when instantiating a client.

This specifically adds support for supplying a CA certificate to use
when connecting to a NATS server. I opted not to add client certificate
support for mTLS for now, but it's an easy lift if it should just be
added now.

Signed-off-by: Dan Norris <protochron@users.noreply.github.com>