feat(*)!: Makes version optional (#281)
This PR makes two major changes. The first, and most important, is that
versions are now optional. If no version field is passed, wadm will
automatically generate a ULID to use as its version. This means that
listing versions can be done in order by sorting them lexographically.

Second is an inversion of the delete behavior. Most of the time the
delete endpoint in practice has been used to delete the whole application
and not just a specific version. Now, by default, if you call the delete
endpoint, it will delete all versions. You can still pass a now optional
version to delete a single specific version.

While I was in the handlers, I also removed an unused field for
orphaning undeploys

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>