ref(*)!: Removes custom mirror code
This removes our custom mirror code and replaces it with multiple filter
subjects. While I was in here, I also changed our other `wadm.evt`
topics to publish their event type in the topic as well. This is marked
as a breaking change because it removes something from the crate API.
This shouldn't affect normal users

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>