fix(chart): reference the correct value for the Jetstream domain
Pull the `jetstreamDomain` value from `config.wadm.nats.jetstreamDomain`
instead of from `config.wadm.jetstreamDomain` since that is what we
define in the values file. It is also a more logical way to group the
value than what the chart was expecting.

Also bump the default version of wadm to the latest one.

Signed-off-by: Dan Norris <protochron@users.noreply.github.com>